requirements.txt
fastapi~=0.115.11
uvicorn~=0.34.0
langchain-community
openai~=1.66.3
faiss-cpu
pydantic~=2.10.6
gTTS~=2.5.4
python-dotenv~=1.0.1
python-multipart
requests~=2.32.3
langchain~=0.3.20
langchain_openai


langgraph~=0.3.18
